💚 better thebe provider positioning, tested on multi project site
