+++ packages/site/src/pages/Root.tsx
@@ -1,12 +1,6 @@
 import type { SiteManifest } from 'myst-config';
 import type { SiteLoader } from '@myst-theme/common';
-import {
-  BaseUrlProvider,
-  ProjectProvider,
-  SiteProvider,
-  Theme,
-  ThemeProvider,
-} from '@myst-theme/providers';
+import { BaseUrlProvider, SiteProvider, Theme, ThemeProvider } from '@myst-theme/providers';
 import {
   Links,
   LiveReload,
@@ -22,44 +16,9 @@ import { DEFAULT_NAV_HEIGHT, renderers } from '../components/index.js';
 import { Analytics } from '../seo/index.js';
 import { Error404 } from './Error404.js';
 import classNames from 'classnames';
-import { ComputeOptionsProvider, useComputeOptions } from '@myst-theme/jupyter';
-import { ThebeBundleLoaderProvider, ThebeServerProvider } from 'thebe-react';
-
-/**
- * Only because we need to access the `useComputeOptions` hook
- *
- */
-function DocumentComputeBodyWrapper({
-  baseurl,
-  children,
-}: React.PropsWithChildren<{ baseurl?: string }>) {
-  const compute = useComputeOptions();
-  return (
-    <ThebeBundleLoaderProvider
-      loadThebeLite={compute?.thebe?.useJupyterLite ?? false}
-      publicPath={baseurl}
-    >
-      <ThebeServerProvider
-        connect={false}
-        options={compute?.thebe}
-        useBinder={compute?.thebe?.useBinder ?? false}
-        useJupyterLite={compute?.thebe?.useJupyterLite ?? false}
-      >
-        {children}
-      </ThebeServerProvider>
-    </ThebeBundleLoaderProvider>
-  );
-}
-
-export type DocumentFeatures = {
-  figureCompute: boolean;
-  notebookCompute: boolean;
-  launchBinder: boolean;
-};
 
 export function Document({
   children,
-  features,
   scripts,
   theme,
   config,
@@ -69,7 +28,6 @@ export function Document({
   top = DEFAULT_NAV_HEIGHT,
 }: {
   children: React.ReactNode;
-  features?: DocumentFeatures;
   scripts?: React.ReactNode;
   theme: Theme;
   config?: SiteManifest;
@@ -88,12 +46,6 @@ export function Document({
         NavLink: NavLink as any,
       };
 
-  const defaultFeatures = {
-    figureCompute: false,
-    notebookCompute: false,
-    launchBinder: false,
-  };
-
   return (
     <html lang="en" className={classNames(theme)} style={{ scrollPadding: top }}>
       <head>
@@ -110,15 +62,7 @@ export function Document({
       <body className="m-0 transition-colors duration-500 bg-white dark:bg-stone-900">
         <ThemeProvider theme={theme} renderers={renderers} {...links} top={top}>
           <BaseUrlProvider baseurl={baseurl}>
-            <SiteProvider config={config}>
-              <ProjectProvider>
-                <ComputeOptionsProvider features={features ?? defaultFeatures}>
-                  <DocumentComputeBodyWrapper baseurl={baseurl}>
-                    {children}
-                  </DocumentComputeBodyWrapper>
-                </ComputeOptionsProvider>
-              </ProjectProvider>
-            </SiteProvider>
+            <SiteProvider config={config}>{children}</SiteProvider>
           </BaseUrlProvider>
         </ThemeProvider>
         <ScrollRestoration />
