+++ themes/article/app/root.tsx
@@ -58,7 +58,6 @@ export default function AppWithReload() {
     <Document
       theme={theme}
       config={config}
-      features={{ notebookCompute: false, figureCompute: true, launchBinder: true }}
       scripts={MODE === 'static' ? undefined : <ContentReload port={CONTENT_CDN_PORT} />}
       staticBuild={MODE === 'static'}
       baseurl={BASE_URL}
