+++ themes/book/app/routes/$.tsx
@@ -1,5 +1,5 @@
-import type { LinksFunction, LoaderFunction, MetaFunction } from '@remix-run/node';
-import { isFlatSite, type PageLoader } from '@myst-theme/common';
+import { json, type LinksFunction, type LoaderFunction, type MetaFunction } from '@remix-run/node';
+import { getProject, isFlatSite, type PageLoader } from '@myst-theme/common';
 import {
   getMetaTagsForArticle,
   KatexCSS,
@@ -17,10 +17,13 @@ import type { SiteManifest } from 'myst-config';
 import {
   TabStateProvider,
   UiStateProvider,
+  useBaseurl,
   useSiteManifest,
   useThemeTop,
+  ProjectProvider,
 } from '@myst-theme/providers';
 import { MadeWithMyst } from '@myst-theme/icons';
+import { ComputeOptionsProvider, ThebeLoaderAndServer } from '@myst-theme/jupyter';
 
 export const meta: MetaFunction = (args) => {
   const config = args.parentsData?.root?.config as SiteManifest | undefined;
@@ -39,17 +42,20 @@ export const links: LinksFunction = () => [KatexCSS];
 
 export const loader: LoaderFunction = async ({ params, request }) => {
   const [first, second] = new URL(request.url).pathname.slice(1).split('/');
-  const project = second ? first : undefined;
+  const projectName = second ? first : undefined;
   const slug = second || first;
   const config = await getConfig();
+  const project = getProject(config, projectName ?? slug);
   const flat = isFlatSite(config);
   const page = await getPage(request, {
-    project: flat ? project : project ?? slug,
-    slug: flat ? slug : project ? slug : undefined,
+    project: flat ? projectName : projectName ?? slug,
+    slug: flat ? slug : projectName ? slug : undefined,
     redirect: process.env.MODE === 'static' ? false : true,
   });
-  return page;
+  return json({ page, project });
 };
+
+type ManifestProject = Required<SiteManifest>['projects'][0];
 
 export function ArticlePageAndNavigation({
   children,
@@ -96,8 +102,10 @@ interface BookThemeTemplateOptions {
 export default function Page() {
   const { container, outline } = useOutlineHeight();
   const top = useThemeTop();
-  const article = useLoaderData<PageLoader>() as PageLoader;
-  const pageDesign: BookThemeTemplateOptions = (article.frontmatter as any)?.options ?? {};
+  const data = useLoaderData() as { page: PageLoader; project: ManifestProject };
+
+  const baseurl = useBaseurl();
+  const pageDesign: BookThemeTemplateOptions = (data.page.frontmatter as any)?.options ?? {};
   const siteDesign: BookThemeTemplateOptions =
     (useSiteManifest() as SiteManifest & BookThemeTemplateOptions)?.options ?? {};
   const { hide_toc, hide_outline, hide_footer_links, outline_maxdepth } = {
@@ -105,20 +113,32 @@ export default function Page() {
     ...pageDesign,
   };
   return (
-    <ArticlePageAndNavigation hide_toc={hide_toc} projectSlug={article.project}>
-      <main ref={container} className="article-grid subgrid-gap col-screen">
-        {!hide_outline && (
-          <div className="sticky z-10 hidden h-0 col-margin-right-inset lg:block" style={{ top }}>
-            <DocumentOutline
-              top={16}
-              className="relative"
-              outlineRef={outline}
-              maxdepth={outline_maxdepth}
-            />
-          </div>
-        )}
-        <ArticlePage article={article} hide_all_footer_links={hide_footer_links} />
-      </main>
+    <ArticlePageAndNavigation hide_toc={hide_toc} projectSlug={data.page.project}>
+      {/* <ProjectProvider project={project}> */}
+      <ProjectProvider>
+        <ComputeOptionsProvider
+          features={{ notebookCompute: true, figureCompute: true, launchBinder: false }}
+        >
+          <ThebeLoaderAndServer baseurl={baseurl}>
+            <main ref={container} className="article-grid subgrid-gap col-screen">
+              {!hide_outline && (
+                <div
+                  className="sticky z-10 hidden h-0 col-margin-right-inset lg:block"
+                  style={{ top }}
+                >
+                  <DocumentOutline
+                    top={16}
+                    className="relative"
+                    outlineRef={outline}
+                    maxdepth={outline_maxdepth}
+                  />
+                </div>
+              )}
+              <ArticlePage article={data.page} hide_all_footer_links={hide_footer_links} />
+            </main>
+          </ThebeLoaderAndServer>
+        </ComputeOptionsProvider>
+      </ProjectProvider>
     </ArticlePageAndNavigation>
   );
 }
